Uniform Buffer Bug Cleanup, Specular lighting (#18)
Summary: After increasing the size of my uniform buffer the program started crashing unexpectedly. Several adjustments were made to set up Uniform Buffers correctly. Namely, persistently mapping the memory vs mapping and unmapping every frame, and ensuring the alignment of the UniformBuffer was correct. However, during some cleanup the program repeatedly crashed, after much investigation and understanding to breakdown the underlying problem, the heap memory that stored the uniform buffer information that was bound to a command buffer was being written to as it was being presented.

The fix: Increasing the total number of uniform buffers to match the number of swapchain images.